Move control of barricades to barricadecontroller
Moved the control of barricades to a seperate barricadeController class.

<interaction, organisation
