+++ src/main/java/spaceinvaders/group_22/MovableUnitController.java
@@ -1,0 +1,9 @@
+package spaceinvaders.group_22;
+/**
+ * Class for controlling movable units.
+ * @author Dorian
+ *
+ */
+public interface MovableUnitController {
+
+}
