+++ src/main/java/spaceinvaders/group_22/SpaceShipController.java
@@ -1,0 +1,8 @@
+package spaceinvaders.group_22;
+/**
+ * Class for controlling the spaceship
+ * @author Dorian
+ */
+public class SpaceShipController extends UnitController implements MovableUnitController {
+
+}
