+++ src/main/java/spaceinvaders/group_22/UnitController.java
@@ -1,0 +1,5 @@
+package spaceinvaders.group_22;
+
+public class UnitController {
+
+}
